roaddialog/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.23
sentence-transformers==2.2.2
redis==5.0.1
bcrypt==4.0.1
jose==1.1.0
python-dotenv==1.0.0
celery==5.3.6
minio==7.2.0